+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The access token in the `Connect-Service` command cannot be within the connection string but is an optional parameter.
 - Connecting to a azure sql server without a access token, will aquire a new token.
 
+### Added
+
+- Added deploy options to `Install-Package`.
+
 ## [1.3.0] - 2023-03-14
 
 ### Added